% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_plus.R
\name{\%+\%}
\alias{\%+\%}
\title{Use the \verb{\%+\%} operator from \code{ggplot2} for \code{ggtibble} and \code{gglist} objects}
\usage{
e1 \%+\% e2
}
\arguments{
\item{e1}{Either a \code{ggtibble} or \code{gglist} object or an object that can use
the default \verb{ggplot2::\%+\%} function}

\item{e2}{A plot component (see ?ggplot2::\verb{\%+\%})}
}
\description{
Use the \verb{\%+\%} operator from \code{ggplot2} for \code{ggtibble} and \code{gglist} objects
}
