% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs_glue.R
\name{labs_glue}
\alias{labs_glue}
\title{Generate ggplot2 labels based on data in a ggtibble}
\usage{
labs_glue(p, ...)
}
\arguments{
\item{p}{The ggtibble object}

\item{...}{Named arguments to be used as \code{ggplot2::labs()} labels where the
value is a glue specification}
}
\value{
\code{p} with the labels modified
}
\description{
Generate ggplot2 labels based on data in a ggtibble
}
