% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglist.R, R/ggtibble.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{vec_arith}
\alias{knit_print}
\alias{group_vars}
\alias{ggplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:group_vars]{group_vars()}}}

  \item{ggplot2}{\code{\link[ggplot2:ggplot]{ggplot()}}}

  \item{knitr}{\code{\link[knitr:knit_print]{knit_print()}}}

  \item{vctrs}{\code{\link[vctrs:vec_arith]{vec_arith()}}}
}}

