% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtibble.R
\name{ggtibble}
\alias{ggtibble}
\alias{ggtibble.data.frame}
\title{Make a tibble where one column is the data to plot, one is the gglist, and
one is the caption}
\usage{
ggtibble(data, ...)

\method{ggtibble}{data.frame}(
  data,
  mapping = ggplot2::aes(),
  ...,
  outercols = group_vars(data),
  labs = list(),
  captionglue = ""
)
}
\arguments{
\item{data}{The data.frame to plot}

\item{...}{Passed to subsequent methods (usually passed to \code{gglist()})}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{outercols}{The columns to have outside the nesting}

\item{labs}{Labels to add via \code{labs_glue()}}

\item{captionglue}{The glue specification for creating the caption}
}
\value{
A data.frame with a column named "data_plot" with the data to plot, "figure" with the gglist, and "caption" with the captions
}
\description{
Make a tibble where one column is the data to plot, one is the gglist, and
one is the caption
}
\section{Methods (by class)}{
\itemize{
\item \code{ggtibble(data.frame)}: The default method for a data.frame or tibble

}}
