% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglist.R
\name{knit_print.gg}
\alias{knit_print.gg}
\title{Print a ggplot (usually within knit_print.gglist)}
\usage{
\method{knit_print}{gg}(
  x,
  ...,
  fig_prefix,
  fig_suffix,
  filename = NULL,
  width = 6,
  height = 4,
  units = "in"
)
}
\arguments{
\item{x}{The gg object (i.e. a ggplot)}

\item{...}{Ignored}

\item{fig_prefix}{Text to \code{cat()} before the figure is printed}

\item{fig_suffix}{Any text to add after the figure}

\item{filename}{A filename saving the plot}

\item{width, height, units}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}
}
\value{
The gg object, invisibly
}
\description{
Print a ggplot (usually within knit_print.gglist)
}
\seealso{
Other knitters: 
\code{\link{knit_print.gglist}()}
}
\concept{knitters}
