% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglist.R
\name{knit_print.gg}
\alias{knit_print.gg}
\title{Print a ggplot (usually within knit_print.gglist)}
\usage{
\method{knit_print}{gg}(
  x,
  ...,
  fig_prefix,
  fig_suffix,
  filename = NULL,
  width = 6,
  height = 4,
  units = "in"
)
}
\arguments{
\item{x}{The gg object (i.e. a ggplot)}

\item{...}{Ignored}

\item{fig_prefix}{Text to \code{cat()} before the figure is printed}

\item{fig_suffix}{Any text to add after the figure.  Defaults to \code{NULL},
which means "auto-select": \code{"\\n\\n\\\\FloatBarrier\\n\\n"} for LaTeX output
when \code{length(x) > float_barrier_after}, otherwise \code{"\\n\\n"}.}

\item{filename}{A filename saving the plot}

\item{width, height}{Plot size in units expressed by the \code{units} argument.
If not supplied, uses the size of the current graphics device.}

\item{units}{One of the following units in which the \code{width} and \code{height}
arguments are expressed: \code{"in"}, \code{"cm"}, \code{"mm"} or \code{"px"}.}
}
\value{
The gg object, invisibly
}
\description{
Print a ggplot (usually within knit_print.gglist)
}
\seealso{
Other knitters:
\code{\link[=knit_print.gglist]{knit_print.gglist()}}
}
\concept{knitters}
