% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglist.R
\name{knit_print.gglist}
\alias{knit_print.gglist}
\title{Print a list of plots made by gglist}
\usage{
knit_print.gglist(x, ..., filename = NULL, fig_suffix = "\\n\\n")
}
\arguments{
\item{x}{The gglist object}

\item{...}{extra arguments to \code{knit_print()}}

\item{filename}{A filename with an optional "\%d" sprintf pattern for saving
the plots}

\item{fig_suffix}{Any text to add after the figure (often
"\n\n\\FloatBarrier\n\n")}
}
\value{
The list, invisibly
}
\description{
Print a list of plots made by gglist
}
\seealso{
Other knitters: 
\code{\link{knit_print.gg}()}
}
\concept{knitters}
