% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglist.R
\name{new_gglist}
\alias{new_gglist}
\title{Create a new \code{gglist} object}
\usage{
new_gglist(x = list())
}
\arguments{
\item{x}{A list of ggplot2 objects to convert into a gglist}
}
\value{
The list verified to be a gglist and with the gglist class
}
\description{
Create a new \code{gglist} object
}
\examples{
new_gglist(list(NULL, ggplot2::ggplot(data = data.frame())))
}
