% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtibble.R
\name{new_ggtibble}
\alias{new_ggtibble}
\title{Create a new \code{ggtibble} object}
\usage{
new_ggtibble(x)
}
\arguments{
\item{x}{A data.frame with a column named "figure" and "caption", and where
the "figure" column is a ggtibble.}
}
\value{
The object with a ggtibble class
}
\description{
Create a new \code{ggtibble} object
}
\examples{
new_ggtibble(tibble::tibble(figure = list(ggplot2::ggplot()), caption = ""))
}
\seealso{
Other New ggtibble objects: 
\code{\link{new_gglist}()}
}
\concept{New ggtibble objects}
